MR-2520 Renamed the getExtension helper
