+++ client/helpers/extractFileExtension.js
